backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
yt-dlp==2024.5.27
groq==0.9.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1